fix: rename the folder from PascalCase to kebap-case to match the convention
re #RI-7091
