+++ redisinsight/ui/src/components/new-index/create-index-step/index.ts
